+++ examples/simple.js
@@ -2,18 +2,7 @@ const { retry, http } = require('../src')
 
 ;(async function () {
   try {
-    await retry(
-      [
-        http('http://localhost:8091', {
-          auth: { user: 'admin', pass: 'changeme' },
-        }),
-      ],
-      {
-        timeout: 5000,
-        verbose: true,
-        interval: 500,
-      }
-    )
+    await retry(http('http://localhost:8080'))
   } catch (err) {
     console.error(err)
   }