+++ spec/index.spec.ts
@@ -2,4 +2,4 @@ require('./file.spec')
 require('./http.spec')
 require('./socket.spec')
 require('./retry.spec')
-
+require('./MultiError.spec')