+++ src/MultiError.ts
@@ -1,3 +1,5 @@
 export default class MultiError extends Error {
-  public errors: any[] = []
+  constructor(message: string | undefined, public errors: any[] = []) {
+    super(message)
+  }
 }